doc: create README
Include description of project and it's major components along with useful links and an outline of what I'll do to build this (will be removed later). Later additions will need to include instructions on how to set up locally and what the UX should be like (including thoughts on UX testing).